move fncc for windows and uefi to ql.os.fncc, next is to move linux
