+++ qiling/os/windows/utils.py
@@ -33,76 +33,6 @@ def read_guid(ql, address):
     return uuid.UUID(bytes_le=bytes(raw_guid))
 
 
-def print_function(ql, passthru, address, function_name, params, ret):
-    function_name = function_name.replace('hook_', '')
-    if function_name in ("__stdio_common_vfprintf", "__stdio_common_vfwprintf",
-                         "printf", "wsprintfW", "sprintf"):
-        return
-    log = '0x%0.2x: %s(' % (address, function_name)
-    for each in params:
-        value = params[each]
-        if type(value) == str or type(value) == bytearray:
-            log += '%s = "%s", ' % (each, value)
-        elif type(value) == tuple:
-            log += '%s = 0x%x, ' % (each, value[0])
-        else:
-            log += '%s = 0x%x, ' % (each, value)
-    log = log.strip(", ")
-    log += ')'
-    if ret is not None:
-        log += ' = 0x%x' % ret
-
-    if passthru:
-        log += ' (PASSTHRU)'
-
-    if ql.output != QL_OUTPUT.DEBUG:
-        log = log.partition(" ")[-1]
-        ql.log.info(log)
-    else:
-        ql.log.debug(log)
-
-
-def read_wstring(ql, address):
-    result = ""
-    char = ql.mem.read(address, 2)
-    # decode as utf-16-le only
-    while char.decode('utf-16-le', errors="backslashreplace") != "\x00":
-        address += 2
-        result += char.decode('utf-16-le', errors="backslashreplace")
-        char = ql.mem.read(address, 2)
-
-    # We need to remove \x00 inside the string. Compares do not work otherwise
-    result = result.replace("\x00", "")
-    string_appearance(ql, result)
-    return result
-
-
-def read_wchar(ql, val):
-    return val.decode('utf-16-le', errors="backslashreplace")
-
-
-def read_punicode_string(ql, address):
-    if ql.archbit == 64:  # Win64
-        ub = ql.mem.read(address, getsizeof(UNICODE_STRING64))
-        us = UNICODE_STRING64.from_buffer_copy(ub)
-    else:  # Win32
-        ub = ql.mem.read(address, getsizeof(UNICODE_STRING32))
-        us = UNICODE_STRING32.from_buffer_copy(ub)
-
-    return read_wstring(ql, us.Buffer)
-
-
-def read_cstring(ql, address):
-    result = ""
-    char = ql.mem.read(address, 1)
-    while char.decode(errors="ignore") != "\x00":
-        address += 1
-        result += char.decode(errors="ignore")
-        char = ql.mem.read(address, 1)
-    string_appearance(ql, result)
-    return result
-
-
 def env_dict_to_array(env_dict):
     env_list = []
     for item in env_dict:
